Update README.md
Description of all possible explorable avenues, team members will now volunteer for the areas they wish to explore and report on in 1 week's time 
by adding your name next to a particular folder. Research independently and hmu for any doubts. Reach out to me when you reach something presentable to be thought of for implementation